improvements-to-get-higher-pana-score/renaming
