+++ lib/src/presentation/pages/widgets/appbar_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:suggest_a_feature/src/presentation/pages/theme/suggestions_theme.dart';
-import 'package:suggest_a_feature/src/presentation/pages/widgets/back_button.dart';
+import 'package:suggest_a_feature/src/presentation/pages/widgets/suggestions_back_button.dart';
 import 'package:suggest_a_feature/src/presentation/utils/dimensions.dart';
 
 class SuggestionsAppBar extends StatelessWidget implements PreferredSizeWidget {
