+++ lib/src/presentation/pages/widgets/suggestions_switch.dart
