+++ lib/src/presentation/pages/widgets/zoomable_image.dart
